BUG: Adding a <Link> causes client side infinite redirect
To reproduce, run the app with this change and watch the URL bar. In Chrome you
won't be able to close the browser tab. If you delete the <Link> line, the hot
code reload feature will kick in and replace the code and you'll be able to
close your browser tab.